+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 apidoc extension Change Log
 2.1.4 under development
 -----------------------
 
-- no changes in this release.
+- Enh #7, #132: Add support for `@property` and `@method` (samdark)
 
 
 2.1.3 February 12, 2020